Removed run.py. The table creation is now in app.py. This was a pain figure out... I don't think he addresses this in the lesson.

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required
 from models.item import ItemModel
 
 
@@ -16,7 +16,7 @@ class Item(Resource):
             help="Every item needs a store id!"
             )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -39,7 +39,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin priviledge required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
